Refactored structure of the calculator
